Update mistral tests
