+++ .github/workflows/test.yml
@@ -32,3 +32,32 @@ jobs:
         run: |
           cd example-copilot
           poetry run pytest tests
+  mistral-copilot:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.11"]
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+      - name: Install Poetry
+        uses: snok/install-poetry@v1
+        with:
+          version: 1.7.1
+          virtualenvs-create: true
+          virtualenvs-in-project: true
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v3
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          cd mistral-copilot
+          python -m pip install --upgrade pip
+          poetry install
+      - name: Run Pytest
+        env:
+          MISTRAL_API_KEY: ${{ secrets.MISTRAL_API_KEY }}
+        run: |
+          cd mistral-copilot
+          poetry run pytest tests
