--- common/common/__init__.py
@@ -1,11 +1,0 @@
-def capture_stream_response(event_stream: str):
-    captured_stream = ""
-    lines = event_stream.split("\n")
-    for line in lines:
-        if line.startswith("event:"):
-            event_type = line.split("event:")[1].strip()
-        if event_type == "copilotMessageChunk" and line.startswith("data:"):
-            data_payload = line.split("data:")[1].strip()
-            data_dict_ = literal_eval(data_payload)
-            captured_stream += data_dict_["delta"]
-    return captured_stream
