+++ mistral-copilot/pyproject.toml
@@ -18,6 +18,7 @@ litellm = "^1.48.19"
 openai = "^1.37.0"
 sse-starlette = "^2.1.2"
 aiohttp = "^3.10.9"
+common = {path = "../common", develop = true}
 
 
 [tool.poetry.group.development.dependencies]
