+++ mistral-copilot/tests/test_copilot.py
@@ -5,6 +5,8 @@ from fastapi.testclient import TestClient
 from mistral_copilot.main import app
 import pytest
 
+from common.testing import capture_stream_response
+
 test_client = TestClient(app)
 
 
@@ -20,34 +22,13 @@ def reset_sse_starlette_appstatus_event():
     AppStatus.should_exit_event = None
 
 
-def _capture_stream_response(event_stream: str) -> tuple[str, str]:
-    if "copilotFunctionCall" in event_stream:
-        event_name = "copilotFunctionCall"
-        event_stream = event_stream.split("\n")
-        data_payload = event_stream[1].split("data:")[-1].strip()
-        return event_name, data_payload
-
-    captured_stream = ""
-    event_name = ""
-    lines = event_stream.split("\n")
-    for line in lines:
-        if line.startswith("event:"):
-            event_type = line.split("event:")[1].strip()
-        if event_type == "copilotMessageChunk" and line.startswith("data:"):
-            event_name = "copilotMessageChunk"
-            data_payload = line.split("data:")[1].strip()
-            data_dict_ = literal_eval(data_payload)
-            captured_stream += data_dict_["delta"]
-    return event_name, captured_stream
-
-
 def test_query():
     test_payload_path = (
         Path(__file__).parent.parent.parent / "test_payloads" / "single_message.json"
     )
     test_payload = json.load(open(test_payload_path))
     response = test_client.post("/v1/query", json=test_payload)
-    event_name, captured_stream = _capture_stream_response(response.text)
+    event_name, captured_stream = capture_stream_response(response.text)
     assert response.status_code == 200
     assert event_name == "copilotMessageChunk"
     assert "2" in captured_stream
@@ -59,7 +40,7 @@ def test_query_conversation():
     )
     test_payload = json.load(open(test_payload_path))
     response = test_client.post("/v1/query", json=test_payload)
-    event_name, captured_stream = _capture_stream_response(response.text)
+    event_name, captured_stream = capture_stream_response(response.text)
     assert response.status_code == 200
     assert event_name == "copilotMessageChunk"
     assert "4" in captured_stream
@@ -73,7 +54,7 @@ def test_query_with_context():
     )
     test_payload = json.load(open(test_payload_path))
     response = test_client.post("/v1/query", json=test_payload)
-    event_name, captured_stream = _capture_stream_response(response.text)
+    event_name, captured_stream = capture_stream_response(response.text)
     assert response.status_code == 200
     assert event_name == "copilotMessageChunk"
     assert "pizza" in captured_stream.lower()
@@ -88,24 +69,34 @@ def test_query_no_messages():
 
 
 def test_query_function_call():
-    test_payload = {
-        "messages": [{"role": "human", "content": "what is the weather in London?"}],
-        "widgets": [
-            {
-                "uuid": "ff6368ec-a397-4baf-9f5a-fecd9fd797a3",
-                "name": "get_weather",
-                "description": "Get the weather for a location",
-            }
-        ],
-    }
+    test_payload_path = (
+        Path(__file__).parent.parent.parent
+        / "test_payloads"
+        / "retrieve_widget_from_dashboard.json"
+    )
+    test_payload = json.load(open(test_payload_path))
     response = test_client.post("/v1/query", json=test_payload)
-    event_name, captured_stream = _capture_stream_response(response.text)
+    event_name, captured_stream = capture_stream_response(response.text)
 
     function_call = literal_eval(captured_stream)
-
     assert response.status_code == 200
     assert event_name == "copilotFunctionCall"
     assert function_call["function"] == "get_widget_data"
     assert function_call["input_arguments"] == {
         "widget_uuid": "ff6368ec-a397-4baf-9f5a-fecd9fd797a3"
     }
+
+
+def test_query_function_call_gives_final_answer():
+    test_payload_path = (
+        Path(__file__).parent.parent.parent
+        / "test_payloads"
+        / "retrieve_widget_from_dashboard_with_result.json"
+    )
+    test_payload = json.load(open(test_payload_path))
+    response = test_client.post("/v1/query", json=test_payload)
+    event_name, captured_stream = capture_stream_response(response.text)
+
+    assert response.status_code == 200
+    assert event_name == "copilotMessageChunk"
+    assert "10 degrees" in captured_stream
